twist_generator: split codes into .hpp & .cpp
