+++ ros/src/computing/perception/localization/packages/twist_generator/CMakeLists.txt
@@ -25,7 +25,7 @@ find_package(catkin REQUIRED
 ## CATKIN_DEPENDS: catkin_packages dependent projects also need
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
-#  INCLUDE_DIRS include
+ INCLUDE_DIRS include
 #  LIBRARIES twist_generator
  CATKIN_DEPENDS 
         autoware_build_flags
@@ -43,8 +43,8 @@ catkin_package(
 ## Specify additional locations of header files
 ## Your package locations should be listed before other locations
 include_directories(
-# include
-${catkin_INCLUDE_DIRS}
+        include
+        ${catkin_INCLUDE_DIRS}
 )
 # 
 
@@ -67,6 +67,8 @@ install(DIRECTORY launch/
         DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch
         PATTERN ".svn" EXCLUDE)
 
+install(DIRECTORY include/twist_generator/
+        DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION})
 
 
 #############
